+++ src/main/ml-modules/root/lib/crudLib.mjs
@@ -5,7 +5,10 @@ import {
   COLLECTION_NAME_USER_PROFILE,
 } from './appConstants.mjs';
 import {
+  CAPABILITY_READ,
   CAPABILITY_UPDATE,
+  ROLE_NAME_MY_COLLECTIONS_DATA_UPDATER,
+  ROLE_NAME_USER_PROFILE_DATA_READER,
   getExclusiveDocumentPermissions,
   getExclusiveRoleNameByUsername,
   throwIfCurrentUserIsServiceAccount,
@@ -35,28 +38,43 @@ import {
   NotFoundError,
 } from './mlErrorsLib.mjs';
 import { getLanguageIdentifier } from './identifierConstants.mjs';
-import { isNonEmptyArray, isDefined, isUndefined } from '../utils/utils.mjs';
+import {
+  getNodeFromObject,
+  isNonEmptyArray,
+  isDefined,
+  isUndefined,
+} from '../utils/utils.mjs';
 
-const DEFAULT_LANG = getLanguageIdentifier('en');
+const DEFAULT_LANG_IRI = getLanguageIdentifier('en');
 const DEFAULT_NEW_DOCUMENT = false;
 const MAX_ATTEMPTS_FOR_NEW_URI = 20;
 
 const DOCUMENT_TYPE_MY_COLLECTION = 'My Collection';
 const DOCUMENT_TYPE_USER_PROFILE = 'User Profile';
 
-function createDocument(docNode, lang = DEFAULT_LANG) {
+function createDocument(docNode, lang = DEFAULT_LANG_IRI) {
   return _insertDocument(docNode, lang, true);
 }
 
-function readDocument(uri, profile = null, lang = 'en') {
+// This function has access to all user profiles and is responsible for restricting access
+// to the name profile when the requesting user is not the owner of the profile.
+function _readDocument(uri, profile = null, lang = 'en') {
   if (fn.docAvailable(uri)) {
-    return applyProfile(cts.doc(uri), profile, lang);
+    const docNode = cts.doc(uri);
+
+    // If a user profile but not the current user's profile, restrict to the name profile.
+    if (isUserProfile(docNode) && new User().getUserIri() !== uri) {
+      profile = 'name';
+    }
+
+    return applyProfile(docNode, profile, lang);
   } else {
     throw new NotFoundError(`Document '${uri}' not found`);
   }
 }
+const readDocument = import.meta.amp(_readDocument);
 
-function updateDocument(docNode, lang = DEFAULT_LANG) {
+function updateDocument(docNode, lang = DEFAULT_LANG_IRI) {
   return _insertDocument(docNode, lang, false);
 }
 
@@ -92,8 +110,8 @@ function _insertDocument(
   throwIfUserIsServiceAccount(user);
 
   // We're to receive the contents of /json but need the full context going forward.
-  readOnlyDocNode = xdmp.toJSON({
-    json: readOnlyDocNode.toObject(),
+  readOnlyDocNode = getNodeFromObject({
+    json: readOnlyDocNode,
   });
 
   let config;
@@ -166,9 +184,10 @@ function _insertDocument(
     }
 
     // Preserve anything outside of /json
-    for (const prop in existingDocNode) {
+    const existingDocObj = existingDocNode.toObject();
+    for (const prop in existingDocObj) {
       if (prop !== 'json') {
-        editableDocObj[prop] = existingDocNode[prop];
+        editableDocObj[prop] = existingDocObj[prop];
       }
     }
 
@@ -218,7 +237,14 @@ function _getUserProfileConfig(
   let docOptions;
   if (newDocument) {
     docOptions = {
-      permissions: getExclusiveDocumentPermissions(user),
+      permissions: getExclusiveDocumentPermissions(user).concat([
+        xdmp.permission(ROLE_NAME_MY_COLLECTIONS_DATA_UPDATER, CAPABILITY_READ),
+        xdmp.permission(
+          ROLE_NAME_MY_COLLECTIONS_DATA_UPDATER,
+          CAPABILITY_UPDATE
+        ),
+        xdmp.permission(ROLE_NAME_USER_PROFILE_DATA_READER, CAPABILITY_READ),
+      ]),
       collections: [
         getTenantRole(),
         COLLECTION_NAME_MY_COLLECTIONS_FEATURE,
@@ -273,7 +299,13 @@ function _getMyCollectionConfig(
   let docOptions;
   if (newDocument) {
     docOptions = {
-      permissions: getExclusiveDocumentPermissions(user),
+      permissions: getExclusiveDocumentPermissions(user).concat([
+        xdmp.permission(ROLE_NAME_MY_COLLECTIONS_DATA_UPDATER, CAPABILITY_READ),
+        xdmp.permission(
+          ROLE_NAME_MY_COLLECTIONS_DATA_UPDATER,
+          CAPABILITY_UPDATE
+        ),
+      ]),
       collections: [
         getTenantRole(),
         COLLECTION_NAME_MY_COLLECTIONS_FEATURE,