+++ src/test/ml-modules/root/test/unitTestConstants.mjs
@@ -2,13 +2,19 @@
  * Constants used exclusively within /src/test
  */
 
+import {
+  IDENTIFIERS,
+  getLanguageIdentifier,
+} from '/lib/identifierConstants.mjs';
+
 const ROLE_NAME_UNIT_TESTER = '%%mlAppName%%-unit-tester';
 const ROLE_NAME_UNIT_TEST_SERVICE_ACCOUNT_READER =
   '%%mlAppName%%-unit-test-service-account-reader';
 const ROLE_NAME_TENANT_ENDPOINT_CONSUMER = '%%mlAppName%%-endpoint-consumer';
 const ROLE_NAME_TENANT_READER = '%%mlAppName%%-reader';
 
-const USERNAME_FOR_REGULAR_USER = '%%mlAppName%%-unit-test-regular-user';
+const USERNAME_FOR_BONNIE = '%%mlAppName%%-unit-test-bonnie';
+const USERNAME_FOR_CLYDE = '%%mlAppName%%-unit-test-clyde';
 const USERNAME_FOR_SERVICE_ACCOUNT = '%%mlAppName%%-unit-test-service-account';
 
 // Filenames are relative to a suite's test-data subdir.
@@ -27,6 +33,7 @@ export {
   ROLE_NAME_TENANT_READER,
   ROLE_NAME_UNIT_TEST_SERVICE_ACCOUNT_READER,
   ROLE_NAME_UNIT_TESTER,
-  USERNAME_FOR_REGULAR_USER,
+  USERNAME_FOR_BONNIE,
+  USERNAME_FOR_CLYDE,
   USERNAME_FOR_SERVICE_ACCOUNT,
 };